Add deploy script for SimpleKitty contracts
