+++ .env.example
@@ -1,5 +1,7 @@
 DEPLOYMENT_ACCOUNT_PRIVATE_KEY=67..14
 DEPLOYMENT_GAS_LIMIT_EXTRA=0.2
-FOREIGN_DEPLOYMENT_GAS_PRICE=1000000000
+HOME_RPC_URL=https://sokol.poa.network
+HOME_DEPLOYMENT_GAS_PRICE=1000000000
 FOREIGN_RPC_URL=https://sokol.poa.network
+FOREIGN_DEPLOYMENT_GAS_PRICE=1000000000
 KITTIES_AMOUNT=1
